Move source types to types.ts to keep consistency
